### 2022-01-11 #4048 @whedon
[PRE REVIEW]: VPsearch: fast exact sequence similarity search for genomic sequences
**Submitting author:** @jvkersch (<a href="http://orcid.org/0000-0002-5813-5659">Joris Vankerschaver</a>)
**Repository:** https://github.com/enthought/vpsearch
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
