### 2022-01-17 #4048 @jvkersch
@danielskatz Thank you for your message. I hope that the tool passes muster, even if it is at the lower end of LOC. It is a full fledged replacement for industry-standard tools after all.

Regarding the DOI, I found the paper in the [ACM digital library](https://dl.acm.org/doi/10.5555/313559.313789) with another DOI ... which is also invalid. I've contacted the ACM digital library, let's see what they suggest.
