### 2022-02-27 #4048 @jvkersch
@lpantano Thanks for the continued effort in dealing with the paper, I really appreciate it. Please let me know if there's anything I can do to facilitate the reviewing process.

Regarding the incorrect DOI that was pointed out above: I actually contacted both ACM and SIAM and while both were really responsive, there's not much they can do at the moment. Apparently it's not clear who has the responsibility for that volume of proceedings, and it doesn't look likely this will be fixed. I would propose just removing the offending DOI.
