### 2022-03-08 #4236 @editorialbot
[REVIEW]: VPsearch: fast exact sequence similarity search for genomic sequences
**Submitting author:** @jvkersch (<a href="http://orcid.org/0000-0002-5813-5659">Joris Vankerschaver</a>)
**Repository:** https://github.com/enthought/vpsearch
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
