+++ paper/paper.bib
@@ -36,8 +36,7 @@
   author = {Yianilos, Peter},
   year = {1993},
   journal = {Fourth Annual ACM-SIAM Symposium on Discrete Algorithms},
-  volume = {93},
-  doi = {10.1145/313559.313789}
+  volume = {93}
 }
 
 @article{2007-wang-NaiveBayesianClassifier,
