### 2022-04-27 62edeca
DOC: Add reference to DIAMOND.
