+++ paper/paper.bib
@@ -163,3 +163,21 @@
   pages={31-39},
   doi={10.1109/MCSE.2010.118}
 }
+
+@article{2021-buchfink-SensitiveProteinAlignments,
+  title = {Sensitive Protein Alignments at Tree-of-Life Scale Using {{DIAMOND}}},
+  author = {Buchfink, Benjamin and Reuter, Klaus and Drost, Hajk-Georg},
+  year = {2021},
+  month = apr,
+  journal = {Nature Methods},
+  volume = {18},
+  number = {4},
+  pages = {366--368},
+  publisher = {{Nature Publishing Group}},
+  issn = {1548-7105},
+  doi = {10.1038/s41592-021-01101-x},
+  abstract = {We are at the beginning of a genomic revolution in which all known species are planned to be sequenced. Accessing such data for comparative analyses is crucial in this new age of data-driven biology. Here, we introduce an improved version of DIAMOND that greatly exceeds previous search performances and harnesses supercomputing to perform tree-of-life scale protein alignments in hours, while matching the sensitivity of the gold standard BLASTP.},
+  copyright = {2021 The Author(s)},
+  langid = {english},
+  keywords = {Computational biology and bioinformatics,Genome informatics,Genomic analysis,Sequencing,Software}
+}
