### 2022-04-28 f6b92f0
MAINT: Make cut-primers script primer-agnostic.
