+++ .github/workflows/build-wheels.yml
@@ -25,7 +25,7 @@ jobs:
       env:
         CIBW_BUILD_VERBOSITY: 1
         CIBW_BEFORE_BUILD: "./ci/build-parasail.sh"
-        CIBW_SKIP: "cp310-macosx_* pp3*"
+        CIBW_SKIP: "pp3*"
         CIBW_ARCHS_MACOS: "x86_64 arm64"
     - uses: actions/upload-artifact@v2
       with:
