### 2022-04-28 de0d136
DEV: Compile universal library on macOS
