+++ .github/workflows/build-wheels.yml
@@ -7,8 +7,14 @@ jobs:
   buildwheels:
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
         python-version: ['3.8']
+      include:
+        - os: ubuntu-latest
+          target: x86_64
+        - os: macos-latest
+          target: x86_64
+        - os: macos-latest
+          target: arm64
 
     runs-on: ${{ matrix.os }}
 
@@ -24,9 +30,9 @@ jobs:
       run: python -m cibuildwheel --output-dir wheelhouse
       env:
         CIBW_BUILD_VERBOSITY: 1
-        CIBW_BEFORE_BUILD: "./ci/build-parasail.sh"
+        CIBW_BEFORE_BUILD: "./ci/build-parasail.sh ${{matrix.target}}"
         CIBW_SKIP: "pp3*"
-        CIBW_ARCHS_MACOS: "x86_64 arm64"
+        CIBW_ARCHS_MACOS: "${{matrix.target}}"
     - uses: actions/upload-artifact@v2
       with:
         path: ./wheelhouse/*.whl
