+++ .github/workflows/build-wheels.yml
@@ -7,10 +7,9 @@ jobs:
   buildwheels:
     strategy:
       matrix:
+        os: [ubuntu-latest, macos-latest]
         python-version: ['3.8']
         include:
-          - os: ubuntu-latest
-            target: x86_64
           - os: macos-latest
             target: x86_64
           - os: macos-latest
