### 2022-04-29 1145e0e
DOC: Add reference to DIAMOND. (#22)
