### 2022-05-01 3894c93
TMP: Install rosetta
