### 2022-05-01 3d7ce15
MAINT: Bump Python version
