+++ .github/workflows/build-wheels.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: ['3.6']
+        python-version: ['3.8']
 
     runs-on: ${{ matrix.os }}
 
