### 2022-05-01 bcde25d
DOC: Update README to mention wheels.
