### 2022-05-01 ac40312
DOC: Add note about runtime/memory usage.
