### 2022-05-04 bbcb6d8
Update install process in README (#25)
* DOC: Update README to mention wheels.

* DOC: Add note about runtime/memory usage.
