### 2022-05-04 041bec6
Apply suggestions from code review
Co-authored-by: Steven Kern <stevenjkern@users.noreply.github.com>
