### 2022-05-04 d09a5b1
STY: Italicize genus and species names.
