### 2022-05-04 8afa4e2
FIX: Add Python/snakemake to environment.
