### 2022-05-04 62c837d
DOC: Clarify that compressed sequence files can also be used.
