### 2022-05-04 3c6d95a
Build macOS wheels for Python 3.10 (#24)
* MAINT: Build macOS wheels

* MAINT: Bump Python version
