### 2022-05-04 10558c8
Update data/README.md
Co-authored-by: Steven Kern <stevenjkern@users.noreply.github.com>
