### 2022-05-06 d68755f
Provide a reproducible example of how to run vpsearch (#21)
* ENH: Reproducible demo database

* DOC: Update README for new data.

* STY: Line breaks

* MAINT: Make cut-primers script primer-agnostic.

* Apply suggestions from code review

Co-authored-by: Steven Kern <stevenjkern@users.noreply.github.com>

* STY: Italicize genus and species names.

* FIX: Add Python/snakemake to environment.

* DOC: Clarify that compressed sequence files can also be used.

* Update data/README.md

Co-authored-by: Steven Kern <stevenjkern@users.noreply.github.com>

Co-authored-by: Steven Kern <stevenjkern@users
[…]
