### 2022-05-07 77ba9a8
DOC: Update acknowledgements to include Peter. (#26)
