### 2022-06-17 #4236 @tamarahadz
Hi @lpantano and @jvkersch I have gone through my checklist and finished my review - After going through the reproducible example on the VPSearch GitHub, package installation was quick and with minimal problems. The package itself was also easy to use and was very quick. I tested it with the reproducible example they provided on their GitHub, and with three fasta files from [Human](https://rnacentral.org/rna/URS000049359E/9606), [Mouse](https://rnacentral.org/rna/URS000058010A/10090) and [E. coli](https://www.ncbi.nlm.nih.gov/nuccore/J01859.1?report=fasta).

Once the issue about community guidelines is resolved, I believe it meets all the requirements on this checklist. I would also add that the README should be reorganized so installation comes before the reproducible example.
