### 2022-07-05 699d816
Guide for contributors (#27)
* DOC: Contributing guide

* DOC: Reorganize README
