### 2022-07-05 #4236 @jvkersch
> Once the issue about community guidelines is resolved, I believe it meets all the requirements on this checklist. I would also add that the README should be reorganized so installation comes before the reproducible example.

I just wanted to add that we addressed the remaining issue on @tamarahadz 's list, and we reorganized the README along the lines suggested here. If anything else is outstanding, please let me know!
