### 2022-07-14 2588a72
DOC: Rename code of conduct
