### 2022-09-07 enthought/vpsearch#30 @sdhutchins
Documentation
Hey, @jvkersch!

I tested the tool out and really like it.

While your readme is very good and extensive, I think it could be enhanced with a couple of more things.

- [x] When trying the tool, I wasn't sure of which all parameters are available. This could be better documented like having the below there. I saw mention of the `-k` parameter as well.

```console
Usage: vpsearch query [OPTIONS] DATABASE QUERY

  Query a built database for sequences.

Options:
  -n, --number INTEGER
  -j, --num-threads INTEGER  Number of threads to use for parallel lookup. The
                             default is to use 1 thread, i.e. to do the lookup
                             serially.
  --help                     Show this message and exit.
```

- [x] Have you all considered using read the docs? https://docs.readthedocs.io/en/stable/tutorial/index.html It may help for some of the details that are in the readme to be there and to give more room to expound.
