### 2022-09-07 enthought/vpsearch#31 @sdhutchins
Great tool!
Hey, @jvkersch! I like this tool a lot. I had to use blast a great deal in the past, and I think this will be very useful once it's published and gains some traction. I'd encourage posting about on https://www.reddit.com/r/bioinformatics/.

Also, I think it'd be helpful to add some badges to the readme and other metrics.

People often look for the below:

- [ ] Code coverage - you already have a great testing infrastructure built so this should be easy.
- [x] Pypi badge - (https://badge.fury.io/for/py/vpsearch)

Neither of those things is required for my review, but I do think they'd help tremendously when people visit the GitHub page to see that.
