### 2022-09-07 032d63e
DOC: Clarify license name
