### 2022-09-07 3a7145c
DOC: Follow 3-part BSD license to the letter.
