### 2022-09-07 #4236 @jvkersch
Thank you @sdhutchins for your review! The points that you raise are valid, and resolving them will improve the tool significantly. I've started addressing the ones that have an immediate solution, but for the ones that require some more work (e.g. switching to RTD) I will need to confer with my coauthors, so this may take a few days.
