### 2022-09-08 9af3eea
Restrict version of setuptools (#35)
setuptools 65.2 and later are incompatible with current EDM Python runtimes. We're looking into the EDM runtime issue, but in the meantime, a workaround is to restrict the version of setuptools used to build the package.
