### 2022-09-08 8b7d390
License terms (#32)
* DOC: Clarify license name

* DOC: Follow 3-part BSD license to the letter.

* MAINT: Update license extension.

* DOC: Reflow lines
