+++ setup.py
@@ -7,21 +7,9 @@ from distutils.core import setup
 from distutils.extension import Extension
 
 import numpy as np
+import parasail
+
 from Cython.Build import cythonize
-
-
-def parasail_get_include():
-    try:
-        return os.environ['PARASAIL_INCLUDE_DIR']
-    except KeyError:
-        return os.path.join(sys.prefix, "include")
-
-
-def parasail_get_lib():
-    try:
-        return os.environ['PARASAIL_LIBRARY_DIR']
-    except KeyError:
-        return os.path.join(sys.prefix, "lib")
 
 
 def get_long_description():
@@ -71,9 +59,9 @@ setup(
             'vpsearch._vpsearch',
             sources=['vpsearch/_vpsearch.pyx'],
             include_dirs=[np.get_include(),
-                          parasail_get_include(),
+                          parasail.get_include(),
                           'vpsearch'],
-            library_dirs=[parasail_get_lib()],
+            library_dirs=[os.path.dirname(parasail.get_library())],
             libraries=['parasail'],
             language='c++',
             extra_compile_args=CPP_BASE_ARGS + ['-std=c++11'],
@@ -92,6 +80,7 @@ setup(
     install_requires=[
         "click",
         "numpy",
+        "parasail",
     ],
     packages=find_packages(),
 )
