### 2022-09-12 0288ec1
DOC: provide documentation, defaults for all options (#33)
