### 2022-09-13 a5a8b0b
MAINT: Make Dockerfile Python-generic
