### 2022-09-24 #4236 @jvkersch
@sdhutchins thanks for the follow-up! There are actually two more things from your review that I wanted to get done: moving the docs to Sphinx (https://github.com/enthought/vpsearch/issues/30) and adding the badges (https://github.com/enthought/vpsearch/issues/31). However, we sort of got onto a side track because we fixed an issue with our build process while addressing your other issue. I can probably take care of the remaining issues in the next week.
