### 2022-09-26 b9270de
Use Parasail from PyPI (#37)
* DEV: Build against PyPI parasail

* MAINT: Remove EDM-specific instructions

* FIX: Do not link against Parasail at build time.

* MAINT: Make Dockerfile Python-generic

* DEP: Pin parasail to 1.3.3
