+++ .github/workflows/build-wheels.yml
@@ -24,7 +24,6 @@ jobs:
       run: python -m cibuildwheel --output-dir wheelhouse
       env:
         CIBW_BUILD_VERBOSITY: 1
-        CIBW_BEFORE_BUILD: "./ci/build-parasail.sh"
         CIBW_SKIP: "pp3*"
     - uses: actions/upload-artifact@v2
       with:
