### 2022-09-26 #4236 @sdhutchins
> @sdhutchins thanks for the follow-up! There are actually two more things from your review that I wanted to get done: moving the docs to Sphinx ([enthought/vpsearch#30](https://github.com/enthought/vpsearch/issues/30)) and adding the badges ([enthought/vpsearch#31](https://github.com/enthought/vpsearch/issues/31)). However, we sort of got onto a side track because we fixed an issue with our build process while addressing your other issue. I can probably take care of the remaining issues in the next week.

Sounds good! Just didn't want to hold up the review for these things, but they definitely would add to your package. Thanks for following up on that @jvkersch.
