+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. VPSearch documentation master file, created by
+   sphinx-quickstart on Tue Sep 27 03:49:01 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to VPSearch's documentation!
+====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
