### 2022-09-26 e94a139
DOC: sphinx-ified documentation
