+++ docs/code_of_conduct.rst
@@ -1,6 +1,11 @@
-# Contributor Covenant Code of Conduct
+.. _coc:
+              
+====================================
+Contributor Covenant Code of Conduct
+====================================
 
-## Our Pledge
+Our Pledge
+~~~~~~~~~~
 
 We as members, contributors, and leaders pledge to make participation in our
 community a harassment-free experience for everyone, regardless of age, body
@@ -12,7 +17,8 @@ identity and orientation.
 We pledge to act and interact in ways that contribute to an open, welcoming,
 diverse, inclusive, and healthy community.
 
-## Our Standards
+Our Standards
+~~~~~~~~~~~~~
 
 Examples of behavior that contributes to a positive environment for our
 community include:
@@ -36,7 +42,8 @@ Examples of unacceptable behavior include:
 * Other conduct which could reasonably be considered inappropriate in a
   professional setting
 
-## Enforcement Responsibilities
+Enforcement Responsibilities
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Community leaders are responsible for clarifying and enforcing our standards of
 acceptable behavior and will take appropriate and fair corrective action in
@@ -48,7 +55,8 @@ comments, commits, code, wiki edits, issues, and other contributions that are
 not aligned to this Code of Conduct, and will communicate reasons for moderation
 decisions when appropriate.
 
-## Scope
+Scope
+~~~~~
 
 This Code of Conduct applies within all community spaces, and also applies when
 an individual is officially representing the community in public spaces.
@@ -56,25 +64,27 @@ Examples of representing our community include using an official e-mail address,
 posting via an official social media account, or acting as an appointed
 representative at an online or offline event.
 
-## Enforcement
+Enforcement
+~~~~~~~~~~~
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
 reported to the community leaders responsible for enforcement:
 
-* [Steven J. Kern](mailto:skern@enthought.com)
-* [Joris Vankerschaver](mailto:joris.vankerschaver@gmail.com)
+* Steven J. Kern (skern@enthought.com)
+* Joris Vankerschaver (joris.vankerschaver@gmail.com)
 
 All complaints will be reviewed and investigated promptly and fairly.
 
 All community leaders are obligated to respect the privacy and security of the
 reporter of any incident.
 
-## Enforcement Guidelines
+Enforcement Guidelines
+~~~~~~~~~~~~~~~~~~~~~~
 
 Community leaders will follow these Community Impact Guidelines in determining
 the consequences for any action they deem in violation of this Code of Conduct:
 
-### 1. Correction
+1. Correction
 
 **Community Impact**: Use of inappropriate language or other behavior deemed
 unprofessional or unwelcome in the community.
@@ -83,7 +93,7 @@ unprofessional or unwelcome in the community.
 clarity around the nature of the violation and an explanation of why the
 behavior was inappropriate. A public apology may be requested.
 
-### 2. Warning
+2. Warning
 
 **Community Impact**: A violation through a single incident or series of
 actions.
@@ -95,7 +105,7 @@ includes avoiding interactions in community spaces as well as external channels
 like social media. Violating these terms may lead to a temporary or permanent
 ban.
 
-### 3. Temporary Ban
+3. Temporary Ban
 
 **Community Impact**: A serious violation of community standards, including
 sustained inappropriate behavior.
@@ -106,7 +116,7 @@ private interaction with the people involved, including unsolicited interaction
 with those enforcing the Code of Conduct, is allowed during this period.
 Violating these terms may lead to a permanent ban.
 
-### 4. Permanent Ban
+4. Permanent Ban
 
[…]
