### 2022-09-26 f8f1267
DOC: Clean up README
