### 2022-09-30 79e8cb4
Sphinx documentation (#39)
* DOC: sphinx-quickstart

* DOC: sphinx-ified documentation

* DOC: Clean up README

* DOC: Make sure _static directory exists

* MAINT: Add RTD config file

* DOC: Update documentation link
