+++ docs/index.rst
@@ -1,0 +1,30 @@
+.. VPSearch documentation master file, created by
+   sphinx-quickstart on Tue Sep 27 03:49:01 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to VPSearch's documentation!
+====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   installation.rst
+   usage.rst
+   contributing.rst
+   implementation.rst
+   wheels.rst
+   references.rst
+
+.. toctree::
+   :maxdepth: 1
+
+   code_of_conduct.rst
+              
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
