+++ docs/references.rst
@@ -1,0 +1,20 @@
+==========
+References
+==========
+
+Vantage point trees were introduced in
+
+- Uhlmann, Jeffrey (1991). "Satisfying General Proximity/Similarity Queries
+  with Metric Trees". Information Processing Letters. 40 (4):
+  175–179. doi:10.1016/0020-0190(91)90074-r.
+
+- Yianilos (1993). Data structures and algorithms for nearest neighbor search
+  in general metric spaces (PDF). Fourth annual ACM-SIAM symposium on Discrete
+  algorithms. Society for Industrial and Applied Mathematics Philadelphia, PA,
+  USA. pp. 311–321. pny93.
+
+The Parasail library is described in
+
+- Daily, Jeff. (2016). Parasail: SIMD C library for global, semi-global, and
+  local pairwise sequence alignments. BMC Bioinformatics, 17(1),
+  1-11. doi:10.1186/s12859-016-0930-z

[…]
