### 2022-09-30 enthought/vpsearch#30 @jvkersch
The docs are now up on https://vpsearch.readthedocs.io/en/latest/.

I'll close the issue as fixed, but feel free to reopen if additional info comes up.
