### 2022-10-03 74a4bb0
DEV: Run tests on merging to main branch
