+++ .github/workflows/run-tests.yaml
@@ -1,10 +1,10 @@
 name: Tests
 
 on:
-  - pull_request
-  - push:
-      branches:
-        - master
+  pull_request:
+  push:
+    branches:
+      - master
 
 jobs:
   tests:
