### 2022-10-03 b72bcbd
Add status badges to README (#40)
* DEV: Run tests on merging to main branch

* DOC: Add badges

* FIX: YAML syntax (hopefully)
