+++ .github/workflows/run-tests.yaml
@@ -1,7 +1,10 @@
 name: Tests
 
 on:
-  pull_request
+  pull_request:
+  push:
+    branches:
+      - master
 
 jobs:
   tests:
