### 2022-10-03 enthought/vpsearch#31 @jvkersch
@sdhutchins I've added the badges to the README. I agree that this would be a good to have, to reflect the status of the project in one look. I added the PyPI badge that you mentioned, and also one to show the status of the tests upon pushes to the main branch. I didn't add the coverage badge, as that would (AFAICT) involve another external service, something that I'd like to avoid. I'll close this issue as fixed, but please feel free to reopen if necessary.
