### 2022-10-03 #4236 @jvkersch
@lpantano @sdhutchins I just wanted to let you know that we've addressed the remaining issues (and some more, too, that came up while we were working on the package). Please let me know if there's anything else you'd like us to address, and thanks again for the many insightful (and actionable!) comments.
