### 2022-10-04 #4236 @jvkersch
@lpantano Thanks for the quick follow-up. Please find the requested information below. Please let me know if anything is missing or incomplete.

- The tagged version: https://github.com/enthought/vpsearch/releases/tag/0.1.2-joss
- Zenodo record: https://zenodo.org/record/7142747
- Zenodo DOI: [10.5281/zenodo.7142747](https://doi.org/10.5281/zenodo.7142747)
