### 2022-10-08 #4236 @Kevin-Mattheus-Moerman
I have reviewed the paper and archive, and all seems in order. I will now proceed to accept this submission in JOSS.
