### 2022-10-08 #4236 @Kevin-Mattheus-Moerman
@jvkersch congratulations on this publication in JOSS :partying_face:

@lpantano thanks a lot for your help editing this work!

Special thanks also to the reviewers @tamarahadz and @sdhutchins !!
